added state and router
