+++ packages/common/package.json
@@ -7,9 +7,14 @@
     "build": "rimraf dist && tsc"
   },
   "dependencies": {
+    "@microsoft/teams-js": "^1.10.0",
+    "@types/react-router": "^5.1.16",
+    "@types/react-router-dom": "^5.1.8",
     "react": "^17.0.2",
     "react-native": "^0.64.2",
-    "react-native-web": "^0.17.1"
+    "react-native-web": "^0.17.1",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0"
   },
   "devDependencies": {
     "@types/react-native": "^0.64.13",
